+++ src/helpers/columnDetails.js
@@ -1,6 +1,6 @@
 /* Helper functions for getting information about a column and its data. */
 
-import { ColumnTypes, UNIQUE_OPTIONS_MAX } from "./constants.js";
+import { ColumnTypes, UNIQUE_OPTIONS_MAX } from "../constants.js";
 
 export function isColumnCategorical(state, column) {
   return (state.columnsByDataType[column] === ColumnTypes.CATEGORICAL);
