+++ src/helpers/metrics.js
@@ -1,7 +1,7 @@
 /*
   Functions for logging analytics metrics via Google Analytics and Firehose.
 */
-import { isUserUploadedDataset, getPercentCorrect } from "./redux";
+import { isUserUploadedDataset, getPercentCorrect } from "../redux";
 
 function getModelMetrics(state) {
   const modelMetrics = {};
