+++ src/helpers/navigationValidation.js
