+++ src/helpers/valueConversion.js
@@ -4,7 +4,7 @@
   likewise convert the returned integers back into human-readable strings.
 */
 import { isEmpty, getKeyByValue } from "./utils.js";
-import { getSelectedCategoricalColumns } from "./redux.js";
+import { getSelectedCategoricalColumns } from "../redux.js";
 
 // Take a ML-friendly integer and convert to human-readable string.
 export function getConvertedValueForDisplay(state, rawValue, column) {
